JL8-24 feat:버튼컴포넌트 위치 조정
